+++ service/controller/key/common.go
@@ -25,6 +25,11 @@ const (
 	DefaultPauseTimeBetweenUpdates = "PT15M"
 )
 
+const (
+	// TerminateUnhealthyNodeResyncPeriod defines resync period for the terminateunhealthynode controller
+	TerminateUnhealthyNodeResyncPeriod = time.Minute * 3
+)
+
 // AMI returns the EC2 AMI for the configured region and given version.
 func AMI(region string, release releasev1alpha1.Release) (string, error) {
 	osVersion, err := OSVersion(release)